Restructure project as a django package
