--- .bandit
@@ -1,2 +1,0 @@
-[bandit]
-exclude = user/tests
