+++ .dockerignore
@@ -6,10 +6,10 @@
 .docker
 
 # Python
-app/__pycache__/
-app/*/__pycache__/
-app/*/*/__pycache__/
-app/*/*/*/__pycache__/
+collectivo-test-app/__pycache__/
+collectivo-test-app/*/__pycache__/
+collectivo-test-app/*/*/__pycache__/
+collectivo-test-app/*/*/*/__pycache__/
 .env/
 .venv/
 venv/
