+++ README.md
@@ -1,6 +1,6 @@
 # collectivo
 
-A dockerized Django REST API with a PostgreSQL database, using [collectivo-ux](https://github.com/MILA-Wien/collectivo-ux/) as a frontend.
+A Django package. Using [collectivo-ux](https://github.com/MILA-Wien/collectivo-ux/) as a frontend.
 
 ## Development
 
@@ -12,12 +12,14 @@ Requirements for the development environment that is used in our team:
 To run and test the app with docker:
 
 - To build a development server, run: `docker compose build`
-- To run a development server, run: `docker compose up -d`
-- To perform tests and linting, run: `docker compose run --rm app sh -c "python manage.py test && flake8"`
+- To run a development server:
+    - If you want to develop just the backend, run: `docker compose up -d`
+    - If you also want to develop the frontend, run `docker compose up -d collectivo db` and follow the instructions at [collectivo-ux](https://github.com/MILA-Wien/collectivo-ux/) to set up a development server for the frontend.
+- To perform tests and linting, run: `docker compose run --rm collectivo sh -c "python manage.py test && flake8"`
 
 ## Documentation
 
 - The API will be available at http://127.0.0.1:8000/.
 - The API documentation will be available at `/api/docs/`.
-- To test the frontend, you can use the API call `GET` at `/api/core/version/`.
+- To test the frontend, you can use the API call `GET` at `/api/version/`.
 - To test backend and frontend together, see [machmit](https://github.com/MILA-Wien/machmit).
