--- app/core/__init__.py
@@ -1,1 +1,0 @@
-"""__init file for core module."""
