--- app/core/urls.py
@@ -1,11 +1,0 @@
-"""Urls contains the url patterns for the core app."""
-from django.urls import path
-
-from core import views
-
-
-app_name = 'core'
-
-urlpatterns = [
-    path('version/', views.GetVersion.as_view(), name='version'),
-]
