+++ collectivo-test-app/.bandit
@@ -1,0 +1,2 @@
+[bandit]
+exclude = collectivo/user/tests
