+++ collectivo-test-app/.flake8
@@ -4,6 +4,3 @@ exclude =
     __pycache__,
     manage.py,
     settings.py,
-
-[bandit]
-exclude = tests
