+++ collectivo-test-app/collectivo-test-app/urls.py
@@ -14,20 +14,11 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from drf_spectacular.views import (
-    SpectacularAPIView,
-    SpectacularSwaggerView
-)
+
 from django.contrib import admin
 from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/core/', include('core.urls')),
-    path('api/user/', include('user.urls')),
-
-    # API Documentation
-    path('api/schema/', SpectacularAPIView.as_view(), name='api-schema'),
-    path('api/docs/', SpectacularSwaggerView.as_view(url_name='api-schema'),
-         name='api-docs'),
+    path("", include('collectivo.urls'))
 ]
