+++ collectivo/__init__.py
@@ -1,0 +1,1 @@
+"""Core module of the collectivo package."""
