+++ collectivo/admin.py
@@ -1,0 +1,4 @@
+"""Admin settings of the collectivo app."""
+# from django.contrib import admin
+
+# Register your models here.
