+++ collectivo/apps.py
@@ -1,0 +1,9 @@
+"""Configuration file for the collectivo app."""
+from django.apps import AppConfig
+
+
+class CollectivoConfig(AppConfig):
+    """Configuration class of the collectivo app."""
+
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'collectivo'
