+++ collectivo/models.py
@@ -1,0 +1,1 @@
+"""Defines the user and user manager model."""
