+++ collectivo/user/models.py
@@ -1,4 +1,4 @@
-"""Defines the user and user manager model."""
+"""Models of the user app."""
 from django.db import models
 from django.contrib.auth.models import (
     AbstractBaseUser,
