+++ collectivo/user/tests/test_user_api.py
@@ -7,9 +7,9 @@ from rest_framework.test import APIClient
 from rest_framework import status
 
 
-CREATE_USER_URL = reverse('user:create')
-TOKEN_URL = reverse('user:token')
-ME_URL = reverse('user:me')
+CREATE_USER_URL = reverse('collectivo:user:create')
+TOKEN_URL = reverse('collectivo:user:token')
+ME_URL = reverse('collectivo:user:me')
 
 
 def create_user(**params):
