+++ collectivo/user/urls.py
@@ -1,6 +1,6 @@
 """URL mappings for the user API."""
 from django.urls import path
-from user import views
+from collectivo.user import views
 
 
 app_name = 'user'
