+++ collectivo/user/views.py
@@ -3,7 +3,7 @@ from rest_framework import generics, authentication, permissions
 from rest_framework.authtoken.views import ObtainAuthToken
 from rest_framework.settings import api_settings
 
-from user.serializers import (
+from collectivo.user.serializers import (
     UserSerializer,
     AuthTokenSerializer
 )
