+++ setup.cfg
@@ -1,0 +1,22 @@
+[metadata]
+name = collectivo
+version = 0.0.1
+classifiers =
+    Environment :: Web Environment
+    Framework :: Django
+    Framework :: Django :: 4.1
+    Operating System :: OS Independent
+    Programming Language :: Python
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Topic :: Internet :: WWW/HTTP
+    Topic :: Internet :: WWW/HTTP :: Dynamic Content
+
+[options]
+include_package_data = true
+packages = collectivo
+python_requires = >=3.8
+install_requires =
+    Django >= 4.1
